Improve model, add more population. Correct serialization of embedded references. Add putative web site.  TODO: Finalize DecoderTyp dependency maintenance as model for other heirarcies; expand lazy children for GET / PUT / POST; make static web site and jsp site work.

+++ src/main/java/com/linepro/modellbahn/model/IDecoderTypCV.java
@@ -14,16 +14,16 @@ public interface IDecoderTypCV {
 
     void setBezeichnung(String bezeichnung);
 
-    Long getMinimal();
+    Integer getMinimal();
 
-    void setMinimal(Long minimal);
+    void setMinimal(Integer minimal);
 
-    Long getMaximal();
+    Integer getMaximal();
 
-    void setMaximal(Long maximal);
+    void setMaximal(Integer maximal);
 
-    Long getWerkseinstellung();
+    Integer getWerkseinstellung();
 
-    void setWerkseinstellung(Long werkseinstellung);
+    void setWerkseinstellung(Integer werkseinstellung);
 
 }
