+++ src/main/java/com/linepro/modellbahn/model/IWahrung.java
@@ -2,8 +2,8 @@ package com.linepro.modellbahn.model;
 
 public interface IWahrung extends INamedItem {
 
-    Long getDecimals();
+    Integer getDecimals();
 
-    void setDecimals(Long decimals);
+    void setDecimals(Integer decimals);
 
 }
